Finish typesafe version prototype of the nRF24 library

+++ firmware/common/common/blt/time.hh
@@ -10,11 +10,13 @@ struct Microseconds {
 
 namespace literals {
 
-constexpr Microseconds operator"" _us(uint64_t us) {
-  return Microseconds{us};
+constexpr Microseconds operator"" _us(uint64_t us)
+{
+  return Microseconds{static_cast<uint32_t>(us)};
 }
-constexpr Microseconds operator"" _ms(uint64_t ms) {
-  return Microseconds{1000 * ms};
+constexpr Microseconds operator"" _ms(uint64_t ms)
+{
+  return Microseconds{1000 * static_cast<uint32_t>(ms)};
 }
 
 }  // namespace literals
